refactor: move examplepicker
