+++ src/web/src/__tests__/components/WSEditorClientConfig.test.tsx
@@ -1,8 +1,8 @@
-import { describe, it, expect, vi, beforeEach } from "vitest";
+import { describe, it, expect, beforeEach, vi } from "vitest";
 import { screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { render } from "../test-utils";
-import WSEditorClientConfigDialog from "../../views/workspace/WSEditorClientConfig";
+import WSEditorClientConfigDialog from "../../views/workspace/WSEditor/WSEditorClientConfig";
 import { workspaceApi, specsApi, errorHandlerApi } from "../../services";
 
 vi.mock("../../services", () => ({
