+++ src/web/src/views/workspace/WSEditor/index.ts
@@ -6,3 +6,4 @@ export { default as WSEditorExportDialog } from "./WSEditorExportDialog";
 export { default as WSEditorDeleteDialog } from "./WSEditorDeleteDialog";
 export { default as WSEditorSwaggerReloadDialog } from "./WSEditorSwaggerReloadDialog";
 export { default as WSRenameDialog } from "./WSRenameDialog";
+export { default as WSEditorClientConfigDialog } from "./WSEditorClientConfig";
